refactor(imports): les 103 imports de domaine visent la racine, les relatifs suivent
`@/@shared/...` et `@/@vitrine/...` deviennent `@/...`. La substitution porte sur des
chemins de modules, pas sur de la prose : elle est mecanique et verifiee par le
compilateur.

Le piege etait ailleurs, dans les chemins relatifs. Le deplacement a change la
profondeur de ContactForm, qui perdait deux niveaux d'un coup (@vitrine et _notPure) :
ses trois imports relatifs pointaient au-dessus de src/. HomeView, lui, importait encore
`components/_notPure/ContactForm`, un segment qui n'existe plus. Quatre lignes corrigees,
et `npx tsc --noEmit` ne rend aucune erreur : aucun module ne reste non resolu.

Biome a reordonne les imports de 23 fichiers, l'ordre alphabetique ayant change avec le
prefixe. Aucune assertion, aucun seuil, aucune regle de lint n'a bouge.

Les .gitkeep des cinq dossiers peuples disparaissent, et les coquilles vides @shared et
@vitrine avec eux.

Refs #143
