+++ src/app/page.tsx
@@ -2,9 +2,9 @@
 // Routage seul : la page d'accueil est composée dans src/@vitrine/views/HomeView.
 
 import type { Metadata } from 'next'
-import { buildPageMetadata } from '@/@shared/seo/page-metadata'
-import { PAGE_ACCUEIL } from '@/@vitrine/contenu/accueil'
-import { HomeView } from '@/@vitrine/views/HomeView'
+import { PAGE_ACCUEIL } from '@/contenu/accueil'
+import { buildPageMetadata } from '@/seo/page-metadata'
+import { HomeView } from '@/views/HomeView'
 
 // Sans cet export, la description écrite pour l'accueil restait du code mort et Google
 // lisait la promesse générique du layout à sa place.
